# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ALMEAA API
  version: 0.4.0
paths:
...
  /api/v1/courses:
    get:
      operationId: contentListCourses
      description: >
        Authenticated staff list. Default limit 50, maximum 100, returns compact rows plus
        hasMore. Teacher scope is forcibly limited to content they own or are explicitly assigned.
        Platform-trainer scope is canonical in Content; an exact pathId+subjectId filter may also
        be authorized by an active Organizations teaching assignment.
      parameters:
        - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { in: query, name: pathId, schema: { type: string } }
        - { in: query, name: subjectId, schema: { type: string } }
        - { in: query, name: search, schema: { type: string, maxLength: 160 } }
        - { in: query, name: workflowStatus, schema: { type: string, enum: [draft, pending_review, approved, rejected, archived] } }
      responses:
        "200": { description: Bounded staff course page }
        "400": { description: Invalid filter or pagination }
        "401": { description: Authentication required }
        "403": { description: Staff scope required }
    post:
      operationId: contentCreateCourse
      description: Platform admins and teachers with canonical platform-trainer scope or an active exact school teaching assignment may create courses. Teacher ownership/assignment is server-forced, revenue share remains admin-controlled, and taxonomy/active Media references are server-validated.
      responses:
        "201": { description: Draft course created }
        "400": { description: Invalid course payload }
        "403": { description: Admin/trainer scope or CSRF denied }
        "409": { description: Taxonomy, owner, asset, or relational conflict }
  /api/v1/courses/{id}:
    get:
      operationId: contentGetCourseStaff
      responses:
        "200": { description: Staff-authorized course detail }
        "403": { description: Ownership or assignment scope denied }
        "404": { description: Course not found }
    put:
      operationId: contentUpdateCourse
      description: CSRF-protected optimistic update using expectedRevision. Approved and archived rows are not edited in place.
      responses:
        "200": { description: Course updated }
        "400": { description: Invalid payload }
        "403": { description: Edit scope or CSRF denied }
        "409": { description: Revision, workflow, taxonomy, owner, or asset conflict }
  /api/v1/courses/{id}/workflow:
    patch:
      operationId: contentSetCourseWorkflow
      description: CSRF-protected workflow transition using expectedRevision. Approval requires publishable course composition.
      responses:
        "200": { description: Course workflow updated }
        "403": { description: Role/scope or CSRF denied }
        "409": { description: Invalid transition or revision conflict }
...
  /api/v1/courses/{id}/modules:
    get:
      operationId: contentListCourseModules
      description: >
        Staff-authorized bounded course composition. At most 200 modules and 5,000 lesson
        placements are returned; placements are loaded without per-module N+1 queries.
      responses:
        "200": { description: Ordered modules and lesson placements }
        "403": { description: Course edit scope denied }
        "404": { description: Course not found }
        "409": { description: Stored composition exceeds safety bounds }
    post:
      operationId: contentCreateCourseModule
      description: >
        CSRF-protected module creation using the parent course expectedRevision.
        Composition changes increment the parent course revision transactionally.
      responses:
        "201": { description: Module created with new courseRevision }
        "400": { description: Invalid module payload }
        "403": { description: Course edit scope or CSRF denied }
        "409": { description: Revision, workflow, or composition bound conflict }
  /api/v1/courses/{id}/modules/{moduleId}:
    put:
      operationId: contentUpdateCourseModule
      description: CSRF-protected optimistic module update/archive; approved or archived courses cannot be composed in place.
      responses:
        "200": { description: Module updated with new courseRevision }
        "400": { description: Invalid module payload }
        "403": { description: Course edit scope or CSRF denied }
        "404": { description: Course or module not found }
        "409": { description: Revision or workflow conflict }
  /api/v1/courses/{id}/modules/{moduleId}/lessons/{lessonId}:
    put:
      operationId: contentPlaceCourseLesson
      description: >
        CSRF-protected optimistic lesson placement. Course and lesson must share taxonomy,
        and trainer callers must have edit authority and active authoring scope for both.
      responses:
        "200": { description: Placement upserted with new courseRevision }
        "400": { description: Invalid placement payload }
        "403": { description: Course/lesson edit scope or CSRF denied }
        "404": { description: Course, module, or lesson not found }
        "409": { description: Revision, taxonomy, workflow, or composition bound conflict }
    delete:
      operationId: contentRemoveCourseLesson
      description: CSRF-protected optimistic removal; request body carries expectedRevision.
      responses:
        "200": { description: Placement removed with new courseRevision }
        "400": { description: Invalid revision payload }
        "403": { description: Course edit scope or CSRF denied }
        "404": { description: Placement not found }
        "409": { description: Revision or workflow conflict }
...
  /api/v1/courses/{id}/publication:
    patch:
      operationId: contentSetCoursePublication
      description: >
        Platform-admin-only, CSRF-protected publication toggle using expectedRevision.
        Publication is separate from review approval and visibility. A course may be published
        only while workflowStatus=approved; leaving approved state clears publication atomically.
      responses:
        "200": { description: Course publication state updated }
        "400": { description: Invalid revision payload }
        "403": { description: Platform admin or CSRF required }
        "404": { description: Course not found }
        "409": { description: Revision or approval-state conflict }
  /api/v1/learning-spaces/{pathId}/subjects/{subjectId}:
    get:
      operationId: contentLearnerLearningSpace
      description: >
        Authenticated learner-safe browsing projection for one active path/subject. Returns
        bounded metadata only: approved+published+visible courses, active+visible Foundation
        topics, and approved+visible library items. It does not resolve entitlement and does
        not expose lesson bodies, file URLs, owner/revenue data, or review metadata.
      parameters:
        - { in: path, name: pathId, required: true, schema: { type: string } }
        - { in: path, name: subjectId, required: true, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 50, default: 24 } }
      responses:
        "200": { description: Bounded learning-space summary with per-section hasMore flags }
        "400": { description: Invalid scope or limit }
        "401": { description: Authentication required }
        "404": { description: Active path/subject scope not found }
  /api/v1/learning-spaces/courses/{id}:
    get:
      operationId: contentLearnerCourseProjection
      description: >
        Authenticated learner-safe course structure for an approved, published and visible
        course. Only active modules and approved+visible lesson metadata are returned.
        Lesson body/media delivery remains outside this endpoint until entitlement resolution exists.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Learner-safe course metadata and module/lesson structure }
        "400": { description: Invalid course ID }
        "401": { description: Authentication required }
        "404": { description: Course unavailable to learner catalog }
        "409": { description: Stored composition exceeds safety bounds }
  /api/v1/learning-spaces/foundation/{id}:
    get:
      operationId: contentLearnerFoundationProjection
      description: >
        Authenticated learner-safe active+visible Foundation topic projection. Linked lessons
        are restricted to approved+visible lesson metadata and linked library items to
        approved+visible metadata. Content/file URLs are not exposed here.
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200": { description: Learner-safe Foundation topic placement metadata }
        "400": { description: Invalid topic ID }
        "401": { description: Authentication required }
        "404": { description: Foundation topic unavailable }
        "409": { description: Stored placement set exceeds safety bounds }